refactor(streaming): archive_counter.py を monthly_archive.py にリネーム (#423)
#156 で `streaming_archive.py` を `daily_archive.py`（日単位）へ移動した結果、
月単位側が汎用名 `archive_counter.py` のまま残り命名対称性が崩れていた。
`monthly_archive.py` にリネームして `daily_archive.py` との姉妹関係を明示する。

- `utils/streaming/archive_counter.py` → `monthly_archive.py`（git mv, move only）
- `tests/test_stream_archive_counter.py` → `test_stream_monthly_archive.py`
- `cli/stream_bandwidth.py` の import を追従
- 公開関数 `count_archives` と `utils/streaming/__init__.py` の公開 API は不変
- ロジック変更・挙動変更なし

Closes #423
